Refactor: Separación de archivos y deshuso de vector por exceso de memoria
