+++ leer_csv.cpp
@@ -4,20 +4,8 @@
 #include <filesystem>
 #include <thread>
 #include <vector>
-
+#include "leer_csv.h"
 using namespace std;
-
-// Estructura para almacenar cada registro del CSV
-struct Registro {
-    string identificador;
-    string especie;
-    string genero;
-    string nombre;
-    string apellido;
-    string fechaNacimiento;
-    string cpOrigen;
-    string cpDestino;
-};
 
 // Obtiene el tamaño en bytes de un archivo dado su ruta
 size_t obtenerTamanoArchivo(const string& rutaArchivo) {
@@ -125,71 +113,4 @@ void procesarBloque(const string& rutaArchivo, size_t inicio, size_t fin, vector
         posActual = archivo.tellg();
         if (archivo.eof()) break;
     }
-}
-
-int main() {
-    // Mostrar el directorio actual
-    cout << "Directorio actual: " << std::filesystem::current_path() << endl;
-
-    string rutaArchivo = "../datos_prueba.csv";
-
-    // Obtener tamaño del archivo
-    size_t tamano = obtenerTamanoArchivo(rutaArchivo);
-    if (tamano == 0) return 1;
-    cout << "Tamano del archivo: " << tamano << " bytes" << endl;
-
-    // Obtener cantidad de hilos
-    unsigned int num_hilos = obtenerCantidadHilos();
-    cout << "Cantidad de hilos detectados: " << num_hilos << endl;
-
-    // Calcular tamaño de bloque
-    size_t tamano_bloque = calcularTamanoBloque(tamano, num_hilos);
-    cout << "Tamano de cada bloque: " << tamano_bloque << " bytes" << endl;
-
-    // 1. Vector para los hilos y para los registros de cada hilo
-    vector<thread> hilos;
-    vector<vector<Registro>> registros_por_hilo(num_hilos);
-
-    // 2. Lanzar los hilos
-    for (unsigned int i = 0; i < num_hilos; ++i) {
-        size_t inicio = i * tamano_bloque;
-        size_t fin = (i == num_hilos - 1) ? tamano : (inicio + tamano_bloque);
-        bool esPrimerHilo = (i == 0);
-        hilos.emplace_back(procesarBloque, rutaArchivo, inicio, fin, ref(registros_por_hilo[i]), esPrimerHilo);
-    }
-
-    // 3. Esperar a que todos los hilos terminen
-    for (auto& h : hilos) {
-        h.join();
-    }
-
-    // 4. (Opcional) Unir todos los registros en un solo vector
-    vector<Registro> todos_los_registros;
-    for (auto& vec : registros_por_hilo) {
-        todos_los_registros.insert(todos_los_registros.end(), vec.begin(), vec.end());
-    }
-
-    cout << "Total de registros leídos: " << todos_los_registros.size() << endl;
-
-    // Comprobación: imprimir la primera y la última línea de datos leídos
-    if (!todos_los_registros.empty()) {
-        cout << "\nPrimer registro:" << endl;
-        cout << "ID: " << todos_los_registros.front().identificador
-             << ", Especie: " << todos_los_registros.front().especie
-             << ", Nombre: " << todos_los_registros.front().nombre << endl;
-
-        cout << "\nÚltimo registro:" << endl;
-        cout << "ID: " << todos_los_registros.back().identificador
-             << ", Especie: " << todos_los_registros.back().especie
-             << ", Nombre: " << todos_los_registros.back().nombre << endl;
-    }
-
-    /* Imprimir la lista de identificadores leídos
-    cout << "\nLista de identificadores leídos:" << endl;
-    for (size_t i = 0; i < todos_los_registros.size(); ++i) {
-        cout << i + 1 << ": " << todos_los_registros[i].identificador << endl;
-    }
-    */
-
-    return 0;
 }
